+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @dexto/core
 
+## 1.6.25
+
 ## 1.6.24
 
 ### Patch Changes