+++ packages/image-local/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @dexto/image-local
 
+## 1.6.25
+
+### Patch Changes
+
+- @dexto/core@1.6.25
+- @dexto/storage@1.6.25
+- @dexto/agent-config@1.6.25
+- @dexto/agent-management@1.6.25
+- @dexto/tools-filesystem@1.6.25
+- @dexto/tools-builtins@1.6.25
+- @dexto/tools-process@1.6.25
+- @dexto/tools-todo@1.6.25
+- @dexto/tools-plan@1.6.25
+- @dexto/tools-scheduler@1.6.25
+- @dexto/tools-lifecycle@1.6.25
+
 ## 1.6.24
 
 ### Patch Changes